backend/requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
charset-normalizer==3.4.4
Django==4.2.7
django-autocomplete-light==3.12.1
django-bootstrap4==25.2
django-bootstrap5==25.2
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.3.0
reportlab==4.4.4
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
python-dotenv
gunicorn
